// src/dungeons_and_trolls_ts/model/dungeonsandtrollsAttributes.ts
/**
 * Dungeons and Trolls
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.10.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class DungeonsandtrollsAttributes {
    'strength'?: number | null;
    'dexterity'?: number | null;
    'intelligence'?: number | null;
    'willpower'?: number | null;
    'constitution'?: number | null;
    'slashResist'?: number | null;
    'pierceResist'?: number | null;
    'fireResist'?: number | null;
    'poisonResist'?: number | null;
    'electricResist'?: number | null;
    'life'?: number | null;
    'stamina'?: number | null;
    'mana'?: number | null;
    'constant'?: number | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "strength",
            "baseName": "strength",
            "type": "number"
        },
        {
            "name": "dexterity",
            "baseName": "dexterity",
            "type": "number"
        },
        {
            "name": "intelligence",
            "baseName": "intelligence",
            "type": "number"
        },
        {
            "name": "willpower",
            "baseName": "willpower",
            "type": "number"
        },
        {
            "name": "constitution",
            "baseName": "constitution",
            "type": "number"
        },
        {
            "name": "slashResist",
            "baseName": "slashResist",
            "type": "number"
        },
        {
            "name": "pierceResist",
            "baseName": "pierceResist",
            "type": "number"
        },
        {
            "name": "fireResist",
            "baseName": "fireResist",
            "type": "number"
        },
        {
            "name": "poisonResist",
            "baseName": "poisonResist",
            "type": "number"
        },
        {
            "name": "electricResist",
            "baseName": "electricResist",
            "type": "number"
        },
        {
            "name": "life",
            "baseName": "life",
            "type": "number"
        },
        {
            "name": "stamina",
            "baseName": "stamina",
            "type": "number"
        },
        {
            "name": "mana",
            "baseName": "mana",
            "type": "number"
        },
        {
            "name": "constant",
            "baseName": "constant",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return DungeonsandtrollsAttributes.attributeTypeMap;
    }
}