// src/dungeons_and_trolls_ts/model/dungeonsandtrollsCharacter.ts
/**
 * Dungeons and Trolls
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.10.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { DungeonsandtrollsAttributes } from './dungeonsandtrollsAttributes';
import { DungeonsandtrollsCoordinates } from './dungeonsandtrollsCoordinates';
import { DungeonsandtrollsEffect } from './dungeonsandtrollsEffect';
import { DungeonsandtrollsItem } from './dungeonsandtrollsItem';
import { DungeonsandtrollsStun } from './dungeonsandtrollsStun';

export class DungeonsandtrollsCharacter {
    'id'?: string;
    'name'?: string;
    'attributes'?: DungeonsandtrollsAttributes;
    'money'?: number;
    'equip'?: Array<DungeonsandtrollsItem>;
    'score'?: number;
    'skillPoints'?: number;
    'effects'?: Array<DungeonsandtrollsEffect>;
    'maxAttributes'?: DungeonsandtrollsAttributes;
    'lastDamageTaken'?: number;
    'coordinates'?: DungeonsandtrollsCoordinates;
    'stun'?: DungeonsandtrollsStun;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "attributes",
            "baseName": "attributes",
            "type": "DungeonsandtrollsAttributes"
        },
        {
            "name": "money",
            "baseName": "money",
            "type": "number"
        },
        {
            "name": "equip",
            "baseName": "equip",
            "type": "Array<DungeonsandtrollsItem>"
        },
        {
            "name": "score",
            "baseName": "score",
            "type": "number"
        },
        {
            "name": "skillPoints",
            "baseName": "skillPoints",
            "type": "number"
        },
        {
            "name": "effects",
            "baseName": "effects",
            "type": "Array<DungeonsandtrollsEffect>"
        },
        {
            "name": "maxAttributes",
            "baseName": "maxAttributes",
            "type": "DungeonsandtrollsAttributes"
        },
        {
            "name": "lastDamageTaken",
            "baseName": "lastDamageTaken",
            "type": "number"
        },
        {
            "name": "coordinates",
            "baseName": "coordinates",
            "type": "DungeonsandtrollsCoordinates"
        },
        {
            "name": "stun",
            "baseName": "stun",
            "type": "DungeonsandtrollsStun"
        }    ];

    static getAttributeTypeMap() {
        return DungeonsandtrollsCharacter.attributeTypeMap;
    }
}