// src/dungeons_and_trolls_ts/model/dungeonsandtrollsCommandsBatch.ts
/**
 * Dungeons and Trolls
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.10.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { DungeonsandtrollsAttributes } from './dungeonsandtrollsAttributes';
import { DungeonsandtrollsIdentifier } from './dungeonsandtrollsIdentifier';
import { DungeonsandtrollsIdentifiers } from './dungeonsandtrollsIdentifiers';
import { DungeonsandtrollsMessage } from './dungeonsandtrollsMessage';
import { DungeonsandtrollsPosition } from './dungeonsandtrollsPosition';
import { DungeonsandtrollsSkillUse } from './dungeonsandtrollsSkillUse';

export class DungeonsandtrollsCommandsBatch {
    'buy'?: DungeonsandtrollsIdentifiers;
    'pickUp'?: DungeonsandtrollsIdentifier;
    'move'?: DungeonsandtrollsPosition;
    'skill'?: DungeonsandtrollsSkillUse;
    'yell'?: DungeonsandtrollsMessage;
    'assignSkillPoints'?: DungeonsandtrollsAttributes;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "buy",
            "baseName": "buy",
            "type": "DungeonsandtrollsIdentifiers"
        },
        {
            "name": "pickUp",
            "baseName": "pickUp",
            "type": "DungeonsandtrollsIdentifier"
        },
        {
            "name": "move",
            "baseName": "move",
            "type": "DungeonsandtrollsPosition"
        },
        {
            "name": "skill",
            "baseName": "skill",
            "type": "DungeonsandtrollsSkillUse"
        },
        {
            "name": "yell",
            "baseName": "yell",
            "type": "DungeonsandtrollsMessage"
        },
        {
            "name": "assignSkillPoints",
            "baseName": "assignSkillPoints",
            "type": "DungeonsandtrollsAttributes"
        }    ];

    static getAttributeTypeMap() {
        return DungeonsandtrollsCommandsBatch.attributeTypeMap;
    }
}