// src/dungeons_and_trolls_ts/model/dungeonsandtrollsCommandsForMonsters.ts
/**
 * Dungeons and Trolls
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.10.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { DungeonsandtrollsCommandsBatch } from './dungeonsandtrollsCommandsBatch';

export class DungeonsandtrollsCommandsForMonsters {
    /**
    * Monster ID with corresponding batch of commands.
    */
    'commands'?: { [key: string]: DungeonsandtrollsCommandsBatch; };

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "commands",
            "baseName": "commands",
            "type": "{ [key: string]: DungeonsandtrollsCommandsBatch; }"
        }    ];

    static getAttributeTypeMap() {
        return DungeonsandtrollsCommandsForMonsters.attributeTypeMap;
    }
}