// src/dungeons_and_trolls_ts/model/dungeonsandtrollsCoordinates.ts
/**
 * Dungeons and Trolls
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.10.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class DungeonsandtrollsCoordinates {
    'level'?: number;
    'positionX'?: number;
    'positionY'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "level",
            "baseName": "level",
            "type": "number"
        },
        {
            "name": "positionX",
            "baseName": "positionX",
            "type": "number"
        },
        {
            "name": "positionY",
            "baseName": "positionY",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return DungeonsandtrollsCoordinates.attributeTypeMap;
    }
}