// src/dungeons_and_trolls_ts/model/dungeonsandtrollsEffect.ts
/**
 * Dungeons and Trolls
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.10.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { DungeonsandtrollsAttributes } from './dungeonsandtrollsAttributes';
import { DungeonsandtrollsDamageType } from './dungeonsandtrollsDamageType';

export class DungeonsandtrollsEffect {
    'damageAmount'?: number;
    'damageType'?: DungeonsandtrollsDamageType;
    'effects'?: DungeonsandtrollsAttributes;
    'duration'?: number;
    'casterId'?: string | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "damageAmount",
            "baseName": "damageAmount",
            "type": "number"
        },
        {
            "name": "damageType",
            "baseName": "damageType",
            "type": "DungeonsandtrollsDamageType"
        },
        {
            "name": "effects",
            "baseName": "effects",
            "type": "DungeonsandtrollsAttributes"
        },
        {
            "name": "duration",
            "baseName": "duration",
            "type": "number"
        },
        {
            "name": "casterId",
            "baseName": "CasterId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return DungeonsandtrollsEffect.attributeTypeMap;
    }
}

export namespace DungeonsandtrollsEffect {
}