// src/dungeons_and_trolls_ts/model/dungeonsandtrollsEvent.ts
/**
 * Dungeons and Trolls
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.10.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { DungeonsandtrollsCoordinates } from './dungeonsandtrollsCoordinates';
import { DungeonsandtrollsEventType } from './dungeonsandtrollsEventType';
import { DungeonsandtrollsSkill } from './dungeonsandtrollsSkill';

export class DungeonsandtrollsEvent {
    'message'?: string;
    'type'?: DungeonsandtrollsEventType;
    'coordinates'?: DungeonsandtrollsCoordinates;
    'damage'?: number | null;
    'playerId'?: string | null;
    'skillName'?: string | null;
    'radius'?: number | null;
    'target'?: DungeonsandtrollsCoordinates;
    'skill'?: DungeonsandtrollsSkill;
    'isRanged'?: boolean | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "DungeonsandtrollsEventType"
        },
        {
            "name": "coordinates",
            "baseName": "coordinates",
            "type": "DungeonsandtrollsCoordinates"
        },
        {
            "name": "damage",
            "baseName": "damage",
            "type": "number"
        },
        {
            "name": "playerId",
            "baseName": "playerId",
            "type": "string"
        },
        {
            "name": "skillName",
            "baseName": "skillName",
            "type": "string"
        },
        {
            "name": "radius",
            "baseName": "radius",
            "type": "number"
        },
        {
            "name": "target",
            "baseName": "target",
            "type": "DungeonsandtrollsCoordinates"
        },
        {
            "name": "skill",
            "baseName": "skill",
            "type": "DungeonsandtrollsSkill"
        },
        {
            "name": "isRanged",
            "baseName": "isRanged",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return DungeonsandtrollsEvent.attributeTypeMap;
    }
}

export namespace DungeonsandtrollsEvent {
}